Add test to ensure that directories are not empty
The scp command will fail if the directory in the preceeding for
loop is empty.  This adds a test to ensure that $i is acutally a
file name.

Closes-Bug: #1621160

+++ multi-node-aio/deploy-vms.sh
@@ -58,7 +58,9 @@ done
 for node in $(get_all_hosts); do
   ssh -q -n -f -o StrictHostKeyChecking=no 10.0.0.${node#*":"} "mkdir -p /tmp/keys"
   for i in /etc/apt/apt.conf.d/00-nokey /etc/apt/sources.list /etc/apt/sources.list.d/* /tmp/keys/*; do
-    scp "$i" "10.0.0.${node#*":"}:$i"
+    if [[ -f "$i" ]]; then
+      scp "$i" "10.0.0.${node#*":"}:$i"
+    fi
   done
   ssh -q -n -f -o StrictHostKeyChecking=no 10.0.0.${node#*":"} "(for i in /tmp/keys/*; do \
       apt-key add \$i; \
